+++ src/components/my-list-button/my-list-button.test.js
@@ -1,0 +1,10 @@
+import MyListButton from "./my-list-button";
+
+it(`MyListButton should render correctly`, () => {
+  const wrapper = renderer
+    .create(
+        <MyListButton movieId={`1`} isFavoriteMovie={true} isAuthenticated={true} onClick={() => {}} />
+    )
+    .toJSON();
+  expect(wrapper).toMatchSnapshot();
+});
